UseLayers hook
