+++ src/types.d.ts
@@ -1,0 +1,5 @@
+
+export type County = {
+  geoid: string;
+  name: string;
+};
